requirements.txt
fastapi==0.111.0
httpx[brotli,http2,socks]==0.28.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
uvicorn[standard]==0.30.1
sentence-transformers==2.6.1
duckduckgo-search==5.3.0
tldextract==5.1.1
trafilatura==1.7.0
feedparser==6.0.12
llama_cpp_python==0.2.76
numpy==1.26.4
pytest-asyncio==0.23.7
